Support compiling with the android.16 profile
Introduces a new build config, DICE_PROFILE_FOR_OPEN_DICE, which
controls the profile_name config that is exported with this module.

Before this change, all trusty modules were being compiled with the
default profile, which was interpretted as "android.14". This change
maintains that behavior by default but allows targets to opt into
the most recent Android profile. This is needed for trusty in
protected VMs.

Bug: 388424977
Test: build.py qemu-generic-arm64-test-debug
Change-Id: I9fefe0a60671b34d1e1d08de191f0c335692566a